+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 ### New Features
 
 ### Bugfixes
-- <a href="https://github.com/openscope/openscope/issues/1565" target="_blank">#1565</a> - Aircraft can no longer taxi to a runway they're already at
+- <a href="https://github.com/openscope/openscope/issues/1565" target="_blank">#1565</a> - Prevent aircraft from taxiing to a runway they're already holding short of
 
 ### Enhancements & Refactors